// src/forum-moderation/enums/forum-moderation-rule-code.enum.ts
export enum ForumModerationRuleCode {
  RATE_LIMIT_OR_DUPLICATE = 'rate_limit_or_duplicate',
  OPENAI_HARASSMENT = 'OPENAI_HARASSMENT',
  OPENAI_HARASSMENT_THREATENING = 'OPENAI_HARASSMENT_THREATENING',

  OPENAI_HATE = 'OPENAI_HATE',
  OPENAI_HATE_THREATENING = 'OPENAI_HATE_THREATENING',

  OPENAI_ILLICIT = 'OPENAI_ILLICIT',
  OPENAI_ILLICIT_VIOLENT = 'OPENAI_ILLICIT_VIOLENT',

  OPENAI_SELF_HARM = 'OPENAI_SELF_HARM',
  OPENAI_SELF_HARM_INTENT = 'OPENAI_SELF_HARM_INTENT',
  OPENAI_SELF_HARM_INSTRUCTIONS = 'OPENAI_SELF_HARM_INSTRUCTIONS',

  OPENAI_SEXUAL = 'OPENAI_SEXUAL',
  OPENAI_SEXUAL_MINORS = 'OPENAI_SEXUAL_MINORS',

  OPENAI_VIOLENCE = 'OPENAI_VIOLENCE',
  OPENAI_VIOLENCE_GRAPHIC = 'OPENAI_VIOLENCE_GRAPHIC',

  OPENAI_MODERATION_FLAGGED = 'OPENAI_MODERATION_FLAGGED',

  LLM_MODERATION_BLOCK = 'LLM_MODERATION_BLOCK',
  TOO_MANY_LINKS = 'too_many_links',
  DUPLICATE_CONTENT = 'duplicate_content',
  REPEATED_CHARACTERS = 'repeated_characters',
  SUSPICIOUS_PROMOTION = 'suspicious_promotion',
  AGGRESSIVE_LANGUAGE = 'aggressive_language',
  TOO_SHORT = 'too_short',
  TOO_LONG = 'too_long',
  UNKNOWN = 'unknown',
}